Moved UploadObservers.java
